# openapi.yaml
openapi: 3.0.3
info:
  title: Byzantium API
  version: 0.1.0
  description: >
    Byzantium is an autonomous-agent trust and compliance gateway. It evaluates
    whether an agent is authorised to perform a given action by enforcing
    operator-defined spend mandates, checking behavioural reputation scores, and
    issuing short-lived ML-DSA–signed PassTokens that payment rails can verify
    in microseconds without calling back to Byzantium. The API also provides
    mandate lifecycle management, liability-receipt recording with Merkle batch
    anchoring, DID-based agent identity, regulator-grade audit export, EIP-3009
    and Solana SPL payment-rail verification, and operator API-key management.
servers:
  - url: http://localhost:8080
    description: Local development server (no TLS)
  - url: https://api.byzantium.example.com
    description: Production — TLS terminated by nginx; same path structure
security:
  - BearerAuth: []
...
paths:
...
  /v1/limits/issue:
    post:
      tags: [Underwriting]
      summary: Underwrite the agent and issue a portable limit
      description: >
        Runs the five controls — standing gate, sublinear growth in settled
        value, experience cap, per-window rate cap, principal consolidation —
        and returns a signed attestation. `reasons` lists every control that
        shaped the number, on an approval as well as a refusal.
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [agent_did]
              properties:
                agent_did: { type: string }
                ccy:
                  type: string
                  enum: [USD, SGD, EUR, GBP, JPY]
                  default: USD
                chains:
                  type: array
                  items: { type: string }
                  description: Chains this limit may be presented on. Empty means any.
                  example: ["base", "solana"]
                asset_classes:
                  type: array
                  items: { type: string, enum: [stablecoin, major_fiat, volatile] }
                counterparty_classes:
                  type: array
                  items: { type: string }
                action_types:
                  type: array
                  items: { type: string }
      responses:
        "200":
          description: Decision. `issued` is false on a refusal, with `refusal` explaining why.
          content:
            application/json:
              schema:
                type: object
                properties:
                  issued: { type: boolean }
                  attestation:
                    $ref: "#/components/schemas/LimitAttestation"
                  tier: { type: string, example: "B2" }
                  score: { type: integer, example: 481 }
                  reasons:
                    type: array
                    items: { type: string }
                  refusal: { type: string }
        "400": { description: No registered principal for this agent }
  /v1/limits/verify:
    post:
      tags: [Underwriting]
      summary: Authorise one draw against a presented limit
      description: >
        The authorisation path. Verifies the signature, checks the credential has
        not been revoked, converts the amount into the attestation's unit of
        account, applies the asset-class haircut, and nets the result against
        exposure recorded across every chain this agent operates on.


        This cannot be done offline: the credential states the ceiling, but only
        the issuer knows how much of it remains, because an agent may be drawing
        on several chains at once and no single venue sees the others.
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [attestation, draw]
              properties:
                attestation:
                  $ref: "#/components/schemas/LimitAttestation"
                draw:
                  type: object
                  required: [amount_minor, currency, asset_class, chain, action_type]
                  properties:
                    amount_minor:
                      type: integer
                      description: Amount in minor units of `currency`.
                    currency: { type: string, enum: [USD, SGD, EUR, GBP, JPY] }
                    asset_class: { type: string, enum: [stablecoin, major_fiat, volatile] }
                    chain: { type: string, example: "solana" }
                    action_type: { type: string, example: "payment" }
                    counterparty_class: { type: string }
                idempotency_key:
                  type: string
                  description: >
                    Supply this and a retry returns the original answer rather
                    than committing the exposure a second time.
      responses:
        "200":
          description: Decision
          content:
            application/json:
              schema:
                type: object
                properties:
                  permitted: { type: boolean }
                  effective_minor:
                    type: integer
                    description: The draw in the attestation's unit of account, after the haircut.
                  effective_ccy: { type: string }
                  window_used_minor: { type: integer }
                  fee_minor: { type: integer }
                  refusal: { type: string }
        "401": { description: Signature invalid, expired, or the credential is revoked }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: >
        Operator API key issued by `POST /v1/keys`. Supply as
        `Authorization: Bearer <key>`. The auth middleware extracts the
        `operator_id` from the key row and injects it into the request context.
...
  schemas:
    LimitAttestation:
      type: object
      description: >
        The portable credential. Short-lived by design: revocation is normally a
        matter of declining to reissue, because per-credential revocation lists
        are the part of every credential system that fails in production.
      properties:
        sub: { type: string, description: Agent DID this limit belongs to. }
        prn: { type: string, description: Pseudonymous principal reference. }
        iss: { type: string, example: "did:web:byzantium" }
        tier: { type: string, example: "B2" }
        lim_single: { $ref: "#/components/schemas/Money" }
        lim_window: { $ref: "#/components/schemas/Money" }
        window_secs: { type: integer }
        ccy: { type: string, enum: [USD, SGD, EUR, GBP, JPY] }
        scope:
          type: object
          properties:
            chains: { type: array, items: { type: string } }
            asset_classes: { type: array, items: { type: string } }
            counterparty_classes: { type: array, items: { type: string } }
            action_types: { type: array, items: { type: string } }
        fee_bps: { type: integer }
        collateral_bps: { type: integer }
        nbf: { type: string, format: date-time }
        exp: { type: string, format: date-time }
        ev:
          type: string
          description: >
            Hash of the evidence behind this decision, never the evidence.
            A dispute is settled with an inclusion proof for the specific event.
        mandate_hash: { type: string }
        guarantee:
          type: object
          nullable: true
          description: Who bears the loss if this limit turns out to be wrong.
          properties:
            model: { type: string, enum: [bureau, underwritten] }
            guarantor: { type: string }
            covered: { $ref: "#/components/schemas/Money" }
        collateral_required:
          nullable: true
          allOf: [{ $ref: "#/components/schemas/Money" }]
        signature: { type: array, items: { type: integer } }
        issuer_pubkey: { type: string }
    Money:
      type: object
      description: >
        An amount in minor units of a currency. Integer-only: limits and exposure
        must never accumulate floating point error, and currencies differ in
        minor-unit scale — JPY has none.
      properties:
        minor_units: { type: integer }
        currency: { type: string, enum: [USD, SGD, EUR, GBP, JPY] }


    # ── Enumerations ──────────────────────────────────────────────────────────
...